Fix output text method and Refactor method and file name
